requirements.txt
Django==3.1
psycopg2-binary==2.8.5
redis>=2.10
celery>=3.1.23
django_celery_beat==2.0.0 
django_tables2==2.3.1 
django-crispy-forms==1.9.2
django-bootstrap-modal-forms==2.0.0
django-widget-tweaks==1.4.8
pytz==2020.1
 